Removed double values

+++ lib/features/game/presentation/widgets/navigation.dart
@@ -28,7 +28,7 @@ class Navigation extends StatelessWidget {
   final VoidCallback breakGame;
   final VoidCallback openRaitingList;
   final num? speed;
-  final double seconds;
+  final int seconds;
 
   @override
   Widget build(BuildContext context) {
